+++ server/model/GameObject.h
@@ -26,7 +26,7 @@ class FloatUpdate;
 
 class GameObject {
     private:
-        const int side;
+        const float side;
 
     protected:
         Position position;
@@ -35,7 +35,7 @@ class GameObject {
         explicit GameObject(const std::vector<float>& position);
         GameObject(const float x, const float y);
         virtual const std::string& get_name() = 0;
-        int get_side();
+        float get_side();
         virtual std::vector<float> get_position();
         bool collided_with(GameObject* other);
         virtual void collide_with(Enemy* enemy) = 0;
